requirements.txt
apache-airflow==2.2.3
scrapy
pymongo==3.11.4
neo4j==4.3.2
